+++ docs/changelog.md
@@ -10,8 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Add autoreconnection to sc-server
 ### Changed
  - Change single logger to a logging hierarchy
-### Removed
- - ScModule, ScAgent and ScKeynode class implementations moved to [py-sc-kpm](https://github.com/ostis-ai/py-sc-kpm/tree/main/src/sc_kpm)
+### Deprecated
+ - ScModule, ScAgent and ScKeynode implementations moved to [py-sc-kpm](https://github.com/ostis-ai/py-sc-kpm)
 
 ## [0.2.4]
 ### Added